+++ CHANGELOG.md
@@ -23,9 +23,9 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   for more information.
 
 ### Removed
-- Drop `TemplateHelper::setCachedConfigurationValue` (#489)
 
 ### Fixed
+- Restore `TemplateHelper::setCachedConfigurationValue` (#527)
 - Fix the seconds per year constant (#520)
 - Harden `unserialize` calls of extension configuration (#510)
 - Stop using things that are deprecated in TYPO3 V9 (#484, #491, #506, #507, #508)